+++ ui/src/components/archive/StoryCard.vue
@@ -552,6 +552,7 @@ p {
   border: 0.5px solid rgba(0,0,0,0.3);
   border-radius: 10px;
   background-color: white;
+  margin-bottom:20px
 }
 
 v-list-item-subtitle {